+++ src/main/java/vn/edu/fpt/petworldplatform/service/CustomerService.java
@@ -140,7 +140,7 @@ public class CustomerService {
         return customerRepo.existsByEmail(email);
     }
 
-    public Optional<Customer> findById(Integer id) {
+    public Optional<Customer> findById(Long id) {
         return customerRepo.findById(id);
     }
